+++ packages/modules/electricity_pricing/flexible_tariffs/smart_energy/config.py
@@ -9,6 +9,6 @@ class SmartEnergyTariffConfiguration:
 @dataclass
 class SmartEnergyTariff:
     name: str = "smartEnergy"
-    type: str = "smartEnergy"
+    type: str = "smart_energy"
     official: bool = True
     configuration: SmartEnergyTariffConfiguration = field(default_factory=SmartEnergyTariffConfiguration)
